feat: divide reducer to separate files
