+++ src/api.js
@@ -1,13 +1,11 @@
 import axios from 'axios';
-import {ActionCreator} from './reducer';
 
 
 /**
  * Api for work with server
- * @param {Function} dispatch
  * @return {AxiosInstance}
  */
-export const createAPI = (dispatch) => {
+export const createAPI = () => {
   const api = axios.create({
     baseURL: `https://es31-server.appspot.com/guess-melody`,
     timeout: 1000 * 5,
@@ -17,10 +15,6 @@ export const createAPI = (dispatch) => {
   const onSuccess = (response) => response;
 
   const onFail = (err) => {
-    if (err.response.status === 403) {
-      dispatch(ActionCreator.requireAuthorization(true));
-    }
-
     throw err;
   };
