+++ src/components/genre-question-screen/genre-question-screen.jsx
@@ -59,9 +59,9 @@ GenreQuestionScreen.propTypes = {
   question: PropTypes.shape({
     answers: PropTypes.arrayOf(PropTypes.shape({
       src: PropTypes.string.isRequired,
-      genre: PropTypes.arrayOf(PropTypes.string).isRequired,
+      genre: PropTypes.string.isRequired,
     })).isRequired,
-    genre: PropTypes.arrayOf(PropTypes.string).isRequired,
+    genre: PropTypes.string.isRequired,
     type: PropTypes.oneOf([`genre`, `artist`]).isRequired,
   }).isRequired,
   renderAnswer: PropTypes.func.isRequired,
